Schedule renderer now handles multiple classes
